flag and other stuff

+++ Gradded-unit/Gradded-unit/src/com/FuturePixels/Drawables/Levels/Flag.java
@@ -97,9 +97,9 @@ public class Flag extends IDrawable {
         if (im instanceof Player && !ran) {
             String Slim = Level().getClass().toString().substring(Level().getClass().toString().lastIndexOf(".") + 1);
             if (!seed.equals("")) {
-                FileUtils.AppendToFile("resources/Savedata/Coop/" + seed + ".txt", "" + String.format("%.2f", Level().getTime()) + "\n");
+                FileUtils.AppendToFile("resources/Savedata/Coop/" + seed + ".txt", "" + String.format("%.2f", (float)Level().getTime()) + "\n");
             } else {
-                FileUtils.AppendToFile("resources/Savedata/" + Slim + ".txt", "" + String.format("%.2f", Level().getTime()) + "\n");
+                FileUtils.AppendToFile("resources/Savedata/" + Slim + ".txt", "" + String.format("%.2f", (float)Level().getTime()) + "\n");
             }
             Level().play("/Sounds/win1.wav");
             Level().play("/sounds/LevelCompleate_" + new Random().nextInt(3) + ".wav");
